+++ cinetica/oscilatorio/__init__.py
@@ -1,1 +1,2 @@
 from .mas import *
+from .mac import *